backend/requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Environment
python-decouple==3.8
python-dotenv==1.0.0

# API Documentation
drf-yasg==1.21.7

# Utilities
pytz==2023.3
Pillow==10.1.0

# Development
django-debug-toolbar==4.2.0
